Work better with Typescript's strict mode
Since users are allowed to pass different bridge types to `AirshipModal`, typing that as `AirshipBridge<unknown>` was actually broken in Typescript's strict mode, but the errors weren't appearing because of a method syntax quirk. Changing the syntax in `AirshipBridge` exposes the errors and makes it straightforward to fix them.

See https://github.com/microsoft/TypeScript/pull/18654